requirements.txt
umap-learn==0.4.1
numpy==1.18.2
pandas==1.0.3
ray==0.8.4
hdbscan==0.8.26
spacy==2.2.4
tweet-preprocessor==0.5.0
tornado==6.0.4
sentence-transformers==0.2.6.1
tweepy==3.8.0
